+++ src/JSystem/J3DGraphLoader/J3DAnmLoader.cpp
@@ -384,10 +384,10 @@ void J3DAnmFullLoader_v15::setAnmVtxColor(J3DAnmVtxColorFull* dst,
 
     for (s32 i = 0; i < dst->mAnmTableNum[0]; i++)
         dst->mAnmVtxColorIndexData[0][i].mpData =
-            (void*)((s32)indexPtr0 + (s32)dst->mAnmVtxColorIndexData[0][i].mpData * 2);
+            (void*)((uintptr_t)indexPtr0 + (uintptr_t)dst->mAnmVtxColorIndexData[0][i].mpData * 2);
     for (s32 i = 0; i < dst->mAnmTableNum[1]; i++)
         dst->mAnmVtxColorIndexData[1][i].mpData =
-            (void*)((s32)indexPtr1 + (s32)dst->mAnmVtxColorIndexData[1][i].mpData * 2);
+            (void*)((uintptr_t)indexPtr1 + (uintptr_t)dst->mAnmVtxColorIndexData[1][i].mpData * 2);
 
     dst->mColorR = JSUConvertOffsetToPtr<u8>(data, data->mRValuesOffset);
     dst->mColorG = JSUConvertOffsetToPtr<u8>(data, data->mGValuesOffset);
@@ -665,10 +665,10 @@ void J3DAnmKeyLoader_v15::setAnmVtxColor(J3DAnmVtxColorKey* dst,
 
     for (s32 i = 0; i < dst->mAnmTableNum[0]; i++)
         dst->mAnmVtxColorIndexData[0][i].mpData =
-            (void*)((s32)indexPtr0 + (s32)dst->mAnmVtxColorIndexData[0][i].mpData * 2);
+            (void*)((uintptr_t)indexPtr0 + (uintptr_t)dst->mAnmVtxColorIndexData[0][i].mpData * 2);
     for (s32 i = 0; i < dst->mAnmTableNum[1]; i++)
         dst->mAnmVtxColorIndexData[1][i].mpData =
-            (void*)((s32)indexPtr1 + (s32)dst->mAnmVtxColorIndexData[1][i].mpData * 2);
+            (void*)((uintptr_t)indexPtr1 + (uintptr_t)dst->mAnmVtxColorIndexData[1][i].mpData * 2);
 
     dst->mColorR = JSUConvertOffsetToPtr<s16>(data, data->mRValOffset);
     dst->mColorG = JSUConvertOffsetToPtr<s16>(data, data->mGValOffset);
